Finished inventory and crafting system
The inventory is more or less done in terms of the base function of storing variables for the user to access. It also has a few handy functions built into script for future ease of use.
The crafting system is functional but a little janky. As long as we don't need to add any mechanics to it we should be fine. There is some necessary hardcoding if we change the number of items that can be crafted together.
Controls so far: M2 to select item to craft, C to craft, V to clear craft.
See notes.txt for more (please actually read it)